add nemotronh-fp8 llama3.2_1b cases
