+++ tests/integration/defs/accuracy/references/gpqa_diamond.yaml
@@ -38,3 +38,7 @@ nvidia/Llama-3_1-Nemotron-Ultra-253B-v1:
   - quant_algo: FP8
     kv_cache_quant_algo: FP8
     accuracy: 57.07
+nvidia/Nemotron-H-8B-Base-8K:
+  - quant_algo: FP8
+    kv_cache_quant_algo: FP8
+    accuracy: 54.50
