+++ tests/integration/test_lists/qa/examples_test_list.txt
@@ -376,15 +376,16 @@ accuracy/test_cli_flow.py::TestLlama3_2_1B::test_weight_streaming[1.0]
 accuracy/test_cli_flow.py::TestLlama3_2_1B::test_cyclic_kv_cache
 accuracy/test_cli_flow.py::TestLlama3_2_1B::test_cyclic_kv_cache_beam_search
 accuracy/test_llm_api_pytorch.py::TestLlama3_2_1B::test_auto_dtype
-accuracy/test_llm_api_pytorch.py::TestLlama3_2_1B::test_smooth_quant
-accuracy/test_llm_api_pytorch.py::TestLlama3_2_1B::test_smooth_quant_ootb
-accuracy/test_llm_api_pytorch.py::TestLlama3_2_1B::test_int4_awq
-accuracy/test_llm_api_pytorch.py::TestLlama3_2_1B::test_int4_awq_int8_kv_cache
 accuracy/test_llm_api_pytorch.py::TestLlama3_2_1B::test_fp8
-accuracy/test_llm_api_pytorch.py::TestLlama3_2_1B::test_fp8_pp2
-accuracy/test_llm_api_pytorch.py::TestLlama3_2_1B::test_fp8_rowwise
-accuracy/test_llm_api_pytorch.py::TestLlama3_2_3B::test_auto_dtype
-accuracy/test_llm_api_pytorch.py::TestLlama3_2_3B::test_fp8_prequantized
+accuracy/test_llm_api.py::TestLlama3_2_1B::test_auto_dtype
+accuracy/test_llm_api.py::TestLlama3_2_1B::test_smooth_quant
+accuracy/test_llm_api.py::TestLlama3_2_1B::test_smooth_quant_ootb
+accuracy/test_llm_api.py::TestLlama3_2_1B::test_int4_awq
+accuracy/test_llm_api.py::TestLlama3_2_1B::test_int4_awq_int8_kv_cache
+accuracy/test_llm_api.py::TestLlama3_2_1B::test_fp8_pp2
+accuracy/test_llm_api.py::TestLlama3_2_1B::test_fp8_rowwise
+accuracy/test_llm_api.py::TestLlama3_2_3B::test_auto_dtype
+accuracy/test_llm_api.py::TestLlama3_2_3B::test_fp8_prequantized
 accuracy/test_cli_flow.py::TestMistral7B::test_beam_search
 accuracy/test_cli_flow.py::TestMistral7B::test_fp8_tp4pp2
 accuracy/test_cli_flow.py::TestMistral7B::test_smooth_quant_tp4pp1
@@ -464,6 +465,10 @@ accuracy/test_llm_api_pytorch.py::TestNemotronNas::test_auto_dtype_tp8
 accuracy/test_llm_api_pytorch.py::TestLlama3_3NemotronSuper49Bv1::test_fp8_prequantized_tp2
 accuracy/test_llm_api_pytorch.py::TestLlama3_1NemotronNano8Bv1::test_fp8_prequantized
 accuracy/test_llm_api_pytorch.py::TestNemotronUltra::test_fp8_prequantized[tp8ep4-cuda_graph=True]
+accuracy/test_cli_flow.py::TestNemotronUltra::test_auto_dtype[tp8-cuda_graph=True] TIMEOUT (240)
+accuracy/test_cli_flow.py::TestNemotronUltra::test_fp8_prequantized[tp8-cuda_graph=True]
+accuracy/test_llm_api_pytorch.py::TestNemotronH::test_auto_dtype
+accuracy/test_llm_api_pytorch.py::TestNemotronH::test_fp8_tp8
 accuracy/test_llm_api_pytorch.py::TestQwen2_7BInstruct::test_auto_dtype
 accuracy/test_llm_api_pytorch.py::TestDeepSeekR1::test_nvfp4_8gpus[latency]
 accuracy/test_llm_api_pytorch.py::TestDeepSeekR1::test_nvfp4_8gpus[latency_trtllmgen]
